+++ core/src/main/java/com/orientechnologies/orient/core/db/OLiveQueryMonitor.java
@@ -7,6 +7,6 @@ public interface OLiveQueryMonitor {
 
   void unSubscribe();
 
-  long getMonitorId();
+  int getMonitorId();
 
 }
